Create Superstore Project
Write a SQL query that orders the items by price.
Write your own SQL query that will show a statistic about the item prices, like a sum, average, minimum, maximum, or count.
Write your own SQL query that will show a statistic about the price for items in the category of "Kitchen Supplies".
NOTE: The values in the categories column are called text strings, and text strings are case sensitive. That means if you want to select "Kitchen Supplies", you will need to type it exactly as it appears in the data, i.e., with a capital "K" and a capital "S".
Come up with your own question about the data and try to answer it using SQL. For example: How many air purifiers are in stock?
It's okay if you can't figure out how to answer your question, the idea here is to practice thinking about what kinds of insights you can get from a dataset from asking questions about it.